requirements.txt
pymilvus==v2.4.4
transformers
einops
timm
pillow
rich
streamlit
streamlit-cropper
